Editor: Lift out the title to its own component
