+++ src/renderer/screens/Editor/Sidebar/Title.js
@@ -1,0 +1,49 @@
+// -*- mode: js-jsx -*-
+/* Chrysalis -- Kaleidoscope Command Center
+ * Copyright (C) 2020  Keyboardio, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from "react";
+import i18n from "i18next";
+
+import Typography from "@material-ui/core/Typography";
+import { withStyles } from "@material-ui/core/styles";
+
+import { KeymapDB } from "../../../../api/keymap";
+
+const styles = theme => ({
+  title: {
+    marginBottom: theme.spacing(2)
+  }
+});
+
+class TitleBase extends React.Component {
+  render() {
+    const { classes, keymap, selectedKey, layer } = this.props;
+
+    const db = new KeymapDB();
+    const label = db.format(keymap.custom[layer][selectedKey], "full");
+
+    return (
+      <Typography variant="h6" className={classes.title}>
+        {label.hint} {label.main} (#{selectedKey}) {" - "}
+        {i18n.t("components.layer", { index: layer })}
+      </Typography>
+    );
+  }
+}
+const Title = withStyles(styles, { withTheme: true })(TitleBase);
+
+export { Title as default };
